router: Cleanup naming
