+++ src/v2/routes.tsx
@@ -67,22 +67,3 @@ export function getAppRoutes(): AppRouteConfig[] {
     { routes: debugRoutes },
   ])
 }
-
-export function getRouteList(): string[] {
-  let flatRoutes: string[] = []
-  const appRoutes = getAppRoutes()[0]
-  if (appRoutes) {
-    appRoutes.children?.forEach(route => {
-      const childRoutes =
-        route.children
-          ?.map(child => child.path)
-          .filter(route => route !== "/" && route !== "*")
-          .map(childPath => route.path + "/" + childPath) || []
-      flatRoutes = flatRoutes.concat(childRoutes).concat(route.path || [])
-    })
-  } else {
-    flatRoutes = []
-  }
-
-  return flatRoutes
-}
