+++ src/v2/server.ts
@@ -2,12 +2,13 @@ import sharify from "sharify"
 import type { NextFunction } from "express"
 import type { ArtsyRequest, ArtsyResponse } from "lib/middleware/artsyExpress"
 import { buildServerApp } from "v2/System/Router/server"
-import { getAppRoutes, getRouteList } from "v2/routes"
+import { getAppRoutes } from "v2/routes"
 import ReactDOM from "react-dom/server"
 import loadAssetManifest from "lib/manifest"
 import express from "express"
 import path from "path"
 import { getServerParam } from "./Utils/getServerParam"
+import { getCacheableRoutes } from "./System/Router/getCacheableRoutes"
 
 // TODO: Use the same variables as the asset middleware. Both config and sharify
 // have a default CDN_URL while this does not.
@@ -22,13 +23,13 @@ if (!NOVO_MANIFEST) {
 
 const app = express()
 const routes = getAppRoutes()
-const routeList = getRouteList()
+const cacheableRoutes = getCacheableRoutes()
 
 /**
  * Mount routes that will connect to global SSR router
  */
 app.get(
-  routeList,
+  cacheableRoutes,
   async (req: ArtsyRequest, res: ArtsyResponse, next: NextFunction) => {
     try {
       const {
